refactor:screen folder
